refactor: split the commands.go grab bag; help covers every command
commands.go had grown to 1120 lines and 37 functions spanning the dev server,
iOS builds, Android builds, tooling install/uninstall, asset generation,
simulator control, plist editing and generic helpers. Split by concern into
tools.go, ios_build.go, android_build.go, build_mobile.go and util.go, leaving
commands.go as dev/serve/test. Pure moves — no logic changed.

Note the filenames: build_ios.go and build_android.go compile only for GOOS=ios
and GOOS=android under Go's implicit build constraints, so both were silently
excluded from the build. Hence ios_build.go / android_build.go.

Six commands (serve, assets, test, version, package, reviews) were dispatched
with no detailed help, and `irgo help help` reported itself unknown. Since the
CLI is the documentation, all of them are covered now.
