+++ cmd/irgo/commands.go
@@ -1,12 +1,9 @@
+// Core commands: the dev server, plain serve, and the test runner.
 package main
 
 import (
 	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
-	"strings"
 )
 
 // runDev starts the development server with hot reload
@@ -52,1069 +49,8 @@ func runServe() error {
 	return fmt.Errorf("no main.go found - are you in an irgo project?")
 }
 
-// runBuild builds for mobile platforms. sim additionally builds the runnable
-// iOS Simulator app (ios target only).
-func runBuild(target string, sim bool) error {
-	// No gomobile check here: buildIOS/buildAndroid both run
-	// ensureMobileBuildSetup, which installs gomobile + gobind when missing.
-	// Gating up front would fail the build before that could happen.
-
-	// Determine module path
-	modulePath, err := getModulePath()
-	if err != nil {
-		return fmt.Errorf("could not determine module path: %w", err)
-	}
-
-	// _templ.go and output.css are generated and gitignored, yet the mobile
-	// package imports the former and every build embeds the latter. Regenerate
-	// both here rather than making callers remember the ordering.
-	if err := ensureAssets(); err != nil {
-		return err
-	}
-
-	// Create build directory
-	if err := os.MkdirAll("build", 0755); err != nil {
-		return fmt.Errorf("creating build directory: %w", err)
-	}
-
-	switch target {
-	case "ios":
-		if err := requireMacOS("iOS"); err != nil {
-			return err
-		}
-		if sim {
-			return buildIOSSim(modulePath)
-		}
-		return buildIOS(modulePath)
-	case "android":
-		return buildAndroid(modulePath)
-	case "all":
-		// Android builds anywhere; iOS cannot leave macOS. Skip rather than
-		// fail so `irgo build all` stays usable on Linux/Windows CI.
-		if runtime.GOOS == "darwin" {
-			if err := buildIOS(modulePath); err != nil {
-				return err
-			}
-		} else {
-			fmt.Printf("Skipping iOS framework: requires macOS (host is %s)\n", runtime.GOOS)
-		}
-		return buildAndroid(modulePath)
-	default:
-		return fmt.Errorf("unknown build target: %s (use ios, android, or all)", target)
-	}
-}
-
-// requireMacOS gates Apple-only targets with an actionable error, instead of
-// letting the build fail later on a missing xcodebuild that the host can never
-// have in the first place.
-func requireMacOS(what string) error {
-	if runtime.GOOS != "darwin" {
-		return fmt.Errorf("%s builds require macOS (Xcode) — cannot cross-compile from %s", what, runtime.GOOS)
-	}
-	return nil
-}
-
-func buildIOS(modulePath string) error {
-	fmt.Println("Building iOS framework...")
-
-	outPath := "build/ios/Irgo.xcframework"
-	if err := os.MkdirAll(filepath.Dir(outPath), 0755); err != nil {
-		return err
-	}
-
-	// Remove existing framework
-	os.RemoveAll(outPath)
-
-	// Ensure go.work and gomobile setup
-	if err := ensureMobileBuildSetup(); err != nil {
-		return fmt.Errorf("mobile build setup failed: %w", err)
-	}
-
-	mobilePackage := modulePath + "/mobile"
-	if err := runGomobileCommand("bind", "-target", "ios", "-o", outPath, mobilePackage); err != nil {
-		return fmt.Errorf("gomobile bind failed: %w", err)
-	}
-	writeArtifactStamp("build/ios")
-
-	fmt.Printf("iOS framework built: %s\n", outPath)
-	return nil
-}
-
-// buildIOSSim builds the runnable simulator app: it scaffolds ios/Example when
-// missing, builds the framework, then drives xcodebuild. Callers (mise tasks,
-// CI) therefore never need to shell out to xcodebuild — or pre-scaffold — for
-// themselves.
-func buildIOSSim(modulePath string) error {
-	if err := checkTool("xcodebuild", "Install Xcode from the App Store"); err != nil {
-		return err
-	}
-	if err := scaffoldExamples(); err != nil {
-		return fmt.Errorf("iOS example scaffold failed: %w", err)
-	}
-	if err := buildIOS(modulePath); err != nil {
-		return err
-	}
-
-	// A previous `irgo run ios --dev` leaves a dev-server URL in Info.plist;
-	// a plain build must produce a self-contained app.
-	clearDevServerInPlist(filepath.Join("ios/Example", "Example/Info.plist"))
-
-	appPath, err := buildSimulatorApp("ios/Example")
-	if err != nil {
-		return err
-	}
-	fmt.Printf("iOS simulator app built: %s\n", appPath)
-	return nil
-}
-
-// buildSimulatorApp runs xcodebuild against the example project and returns the
-// path to the built .app. Shared by `irgo build ios --sim` (build only) and
-// `irgo run ios` (build, then boot/install/launch).
-func buildSimulatorApp(iosProjectPath string) (string, error) {
-	// Find the workspace or project
-	var buildCmd []string
-	// Use generic simulator destination to work with any available iPhone
-	destination := "generic/platform=iOS Simulator"
-	if _, err := os.Stat(filepath.Join(iosProjectPath, "Example.xcworkspace")); err == nil {
-		buildCmd = []string{"xcodebuild", "-workspace", filepath.Join(iosProjectPath, "Example.xcworkspace"),
-			"-scheme", "Example", "-destination", destination,
-			"-derivedDataPath", "build/ios/DerivedData"}
-	} else if _, err := os.Stat(filepath.Join(iosProjectPath, "Example.xcodeproj")); err == nil {
-		buildCmd = []string{"xcodebuild", "-project", filepath.Join(iosProjectPath, "Example.xcodeproj"),
-			"-scheme", "Example", "-destination", destination,
-			"-derivedDataPath", "build/ios/DerivedData"}
-	} else {
-		return "", fmt.Errorf("no Xcode project found in %s", iosProjectPath)
-	}
-
-	fmt.Println("Building iOS app...")
-	if err := runCommand(buildCmd[0], buildCmd[1:]...); err != nil {
-		return "", fmt.Errorf("xcodebuild failed: %w", err)
-	}
-
-	// Find the built app
-	appPath := "build/ios/DerivedData/Build/Products/Debug-iphonesimulator/Example.app"
-	if _, err := os.Stat(appPath); os.IsNotExist(err) {
-		return "", fmt.Errorf("built app not found at %s", appPath)
-	}
-	return appPath, nil
-}
-
-// Artifact version stamps: dev mode reuses previously built frameworks/AARs,
-// but a framework built by an older irgo exposes an older bridge API and the
-// native shells fail to compile against it. The stamp forces a rebuild after
-// an irgo upgrade.
-
-func writeArtifactStamp(dir string) {
-	_ = os.WriteFile(filepath.Join(dir, ".irgo-version"), []byte(version), 0644)
-}
-
-func artifactUpToDate(artifact, stampDir string) bool {
-	if _, err := os.Stat(artifact); err != nil {
-		return false
-	}
-	data, err := os.ReadFile(filepath.Join(stampDir, ".irgo-version"))
-	return err == nil && strings.TrimSpace(string(data)) == version
-}
-
-func buildAndroid(modulePath string) error {
-	fmt.Println("Building Android AAR...")
-
-	outPath := "build/android/irgo.aar"
-	if err := os.MkdirAll(filepath.Dir(outPath), 0755); err != nil {
-		return err
-	}
-
-	// Remove existing AAR
-	os.Remove(outPath)
-
-	// Self-provision the toolchain (JDK/SDK/NDK/gomobile) if anything is
-	// missing — devs and CI never need a separate setup step.
-	if err := ensureAndroidToolchain(false, "irgo"); err != nil {
-		return fmt.Errorf("Android toolchain setup failed: %w", err)
-	}
-
-	// Ensure the canonical android/Example app exists so the AAR has a
-	// consumer out of the box (scaffolded from embedded templates).
-	if err := scaffoldExamples(); err != nil {
-		return fmt.Errorf("Android example scaffold failed: %w", err)
-	}
-
-	// Ensure go.work and gomobile setup
-	if err := ensureMobileBuildSetup(); err != nil {
-		return fmt.Errorf("mobile build setup failed: %w", err)
-	}
-
-	mobilePackage := modulePath + "/mobile"
-	// Pin the Android API level explicitly: gomobile defaults to API 16,
-	// which modern NDKs (r26/r27) reject with "unsupported API version 16".
-	// API 21 is the floor every current NDK supports.
-	if err := runGomobileCommand("bind", "-target", "android", "-androidapi", "21", "-o", outPath, mobilePackage); err != nil {
-		return fmt.Errorf("gomobile bind failed: %w", err)
-	}
-	writeArtifactStamp("build/android")
-
-	fmt.Printf("Android AAR built: %s\n", outPath)
-
-	// Copy to Example project if it exists
-	exampleLibsPath := "android/Example/app/libs/irgo.aar"
-	if _, err := os.Stat("android/Example"); err == nil {
-		os.MkdirAll(filepath.Dir(exampleLibsPath), 0755)
-		if err := copyFile(outPath, exampleLibsPath); err != nil {
-			fmt.Printf("Warning: could not copy to example project: %v\n", err)
-		} else {
-			fmt.Printf("Copied to: %s\n", exampleLibsPath)
-		}
-	}
-
-	return nil
-}
-
-// runTempl generates templ files
-func runTempl() error {
-	if err := ensureGoTool("templ"); err != nil {
-		return err
-	}
-
-	fmt.Println("Generating templ files...")
-	return runCommand("templ", "generate")
-}
-
-// goToolPkg maps a tool name to the module to `go install`. templ is pinned to
-// the project's go.mod version: the generator and the library must match, and
-// @latest drift breaks generated code.
-func goToolPkg(name string) string {
-	switch name {
-	case "templ":
-		if v := templVersionFromGoMod(); v != "" {
-			return "github.com/a-h/templ/cmd/templ@v" + v
-		}
-		return "github.com/a-h/templ/cmd/templ@latest"
-	case "air":
-		return "github.com/air-verse/air@latest"
-	case "gomobile":
-		return "golang.org/x/mobile/cmd/gomobile@latest"
-	case "gobind":
-		return "golang.org/x/mobile/cmd/gobind@latest"
-	}
-	return ""
-}
-
-// irgoToolsDir holds one marker file per tool irgo installed itself, so
-// uninstall can remove exactly those and leave a developer's own copies alone.
-// Without it an uninstall would either delete tools irgo never owned or, if it
-// played safe and skipped them, leave a half-provisioned machine that silently
-// fails to re-provision.
-func irgoToolsDir() string {
-	return filepath.Join(homeDir(), ".irgo", "tools")
-}
-
-func markToolInstalled(name string) {
-	dir := irgoToolsDir()
-	if err := os.MkdirAll(dir, 0o755); err != nil {
-		return
-	}
-	_ = os.WriteFile(filepath.Join(dir, name), []byte("irgo "+version+"\n"), 0o644)
-}
-
-func toolInstalledByIrgo(name string) bool {
-	_, err := os.Stat(filepath.Join(irgoToolsDir(), name))
-	return err == nil
-}
-
-// ensureGoTool installs a Go-based tool when missing rather than printing an
-// install command for someone to copy. `go install` behaves the same on macOS,
-// Linux and Windows, so this needs no per-OS branching.
-func ensureGoTool(name string) error {
-	if _, err := exec.LookPath(name); err == nil {
-		return nil
-	}
-	pkg := goToolPkg(name)
-	if pkg == "" {
-		return fmt.Errorf("%s not found, and no install source is known for it", name)
-	}
-	fmt.Printf("%s not found — installing %s...\n", name, pkg)
-	if err := runCommand(goBin(), "install", pkg); err != nil {
-		return fmt.Errorf("installing %s: %w", name, err)
-	}
-	markToolInstalled(name)
-	// `go install` lands in GOBIN (default $GOPATH/bin), which is often absent
-	// from PATH — prepend it so the tool resolves for the rest of this process.
-	if dir := gobinDir(); dir != "" {
-		_ = os.Setenv("PATH", dir+string(os.PathListSeparator)+os.Getenv("PATH"))
-	}
-	if _, err := exec.LookPath(name); err != nil {
-		return fmt.Errorf("%s still not found after installing it — is %s on your PATH?", name, gobinDir())
-	}
-	return nil
-}
-
-// jsRunner returns the package runner to drive package.json scripts with,
-// preferring bun and falling back to npm. Empty when neither is installed.
-func jsRunner() string {
-	for _, r := range []string{"bun", "npm"} {
-		if _, err := exec.LookPath(r); err == nil {
-			return r
-		}
-	}
-	return ""
-}
-
-// runCSS rebuilds the Tailwind stylesheet. static/css/output.css is generated
-// and gitignored, but it is embedded into every build — so skipping it ships an
-// unstyled app. No-op for projects without a "css" script.
-func runCSS() error {
-	data, err := os.ReadFile("package.json")
-	if err != nil || !strings.Contains(string(data), `"css"`) {
-		return nil // not a Tailwind project
-	}
-	runner := jsRunner()
-	if runner == "" {
-		fmt.Println("  (skipping CSS: neither bun nor npm found)")
-		return nil
-	}
-	if _, err := os.Stat("node_modules"); os.IsNotExist(err) {
-		fmt.Printf("Installing frontend dependencies (%s install)...\n", runner)
-		if err := runCommand(runner, "install"); err != nil {
-			return fmt.Errorf("%s install failed: %w", runner, err)
-		}
-	}
-	fmt.Println("Building CSS...")
-	return runCommand(runner, "run", "css")
-}
-
-// ensureAssets regenerates everything that is gitignored yet embedded into a
-// build: _templ.go and the Tailwind stylesheet. Every build path runs this, so
-// a fresh clone builds correctly without the caller sequencing it by hand.
-func ensureAssets() error {
-	if err := runTempl(); err != nil {
-		return err
-	}
-	return runCSS()
-}
-
 // runTest runs the test suite
 func runTest() error {
 	fmt.Println("Running tests...")
 	return runCommand(goBin(), "test", "-v", "./...")
-}
-
-// installTools installs required development tools
-func installTools() error {
-	fmt.Println("Installing irgo development tools...")
-	fmt.Println()
-
-	// Pin the templ generator to the templ library version in the project's
-	// go.mod (they must stay in sync — @latest drift breaks generated code).
-	templPkg := "github.com/a-h/templ/cmd/templ@latest"
-	if v := templVersionFromGoMod(); v != "" {
-		templPkg = "github.com/a-h/templ/cmd/templ@v" + v
-		fmt.Printf("  templ: pinning to go.mod version v%s\n", v)
-	}
-
-	tools := []struct {
-		name string
-		pkg  string
-	}{
-		{"templ", templPkg},
-		{"air", "github.com/air-verse/air@latest"},
-		{"gomobile", "golang.org/x/mobile/cmd/gomobile@latest"},
-	}
-
-	for _, tool := range tools {
-		if _, err := exec.LookPath(tool.name); err != nil {
-			fmt.Printf("Installing %s...\n", tool.name)
-			cmd := exec.Command(goBin(), "install", tool.pkg)
-			cmd.Stdout = os.Stdout
-			cmd.Stderr = os.Stderr
-			if err := cmd.Run(); err != nil {
-				fmt.Printf("  Warning: failed to install %s: %v\n", tool.name, err)
-			} else {
-				markToolInstalled(tool.name)
-				fmt.Printf("  %s: installed\n", tool.name)
-			}
-		} else {
-			fmt.Printf("  %s: already installed\n", tool.name)
-		}
-	}
-
-	// Initialize gomobile
-	fmt.Println()
-	fmt.Println("Initializing gomobile...")
-	if err := runCommand("gomobile", "init"); err != nil {
-		fmt.Printf("Warning: gomobile init failed: %v\n", err)
-		fmt.Println("You may need to run 'gomobile init' manually after installing Android NDK")
-	}
-
-	fmt.Println()
-	fmt.Println("Tools installed! You may also want to install:")
-	fmt.Println("  - Xcode: from App Store (for iOS development)")
-	fmt.Println("  - Android Studio: https://developer.android.com/studio (for Android development)")
-
-	return nil
-}
-
-// runMobile builds and runs on mobile simulator
-func runMobile(platform string, devMode bool, avdName string, headless bool) error {
-	switch platform {
-	case "ios":
-		return runIOS(devMode)
-	case "android":
-		return runAndroid(devMode, avdName, headless)
-	default:
-		return fmt.Errorf("unknown platform: %s (use ios or android)", platform)
-	}
-}
-
-func runIOS(devMode bool) error {
-	// Check for Xcode
-	if err := checkTool("xcodebuild", "Install Xcode from the App Store"); err != nil {
-		return err
-	}
-	if err := checkTool("xcrun", "Install Xcode Command Line Tools: xcode-select --install"); err != nil {
-		return err
-	}
-
-	// Ensure the canonical ios/Example app exists (scaffolded from the
-	// embedded templates when missing — devs and CI never hand-copy it).
-	if err := scaffoldExamples(); err != nil {
-		return fmt.Errorf("iOS example scaffold failed: %w", err)
-	}
-
-	// Check if ios/Example project exists
-	iosProjectPath := "ios/Example"
-	if _, err := os.Stat(iosProjectPath); os.IsNotExist(err) {
-		return fmt.Errorf("iOS project not found at %s", iosProjectPath)
-	}
-
-	// Dev server URL for simulator to connect to
-	devServerURL := "http://localhost:8080"
-	var devServerCmd *exec.Cmd
-
-	if devMode {
-		fmt.Println("Running in DEV MODE with hot reload...")
-		fmt.Println()
-
-		// Check for required dev tools
-		if err := ensureGoTool("air"); err != nil {
-			return err
-		}
-
-		// Dev mode serves the app from localhost:8080, so a fresh gomobile
-		// build isn't needed. The Xcode project still links against
-		// build/ios/Irgo.xcframework, so build it only when missing or built
-		// by a different irgo version (stale bridge API).
-		if !artifactUpToDate("build/ios/Irgo.xcframework", "build/ios") {
-			modulePath, err := getModulePath()
-			if err != nil {
-				return fmt.Errorf("could not determine module path: %w", err)
-			}
-
-			fmt.Println("Building iOS framework (missing or built by another irgo version)...")
-			if err := buildIOS(modulePath); err != nil {
-				return err
-			}
-		} else {
-			fmt.Println("Using existing build/ios/Irgo.xcframework (delete it to force a rebuild)")
-		}
-
-		// Update Info.plist to enable dev mode
-		infoPlistPath := filepath.Join(iosProjectPath, "Example/Info.plist")
-		if err := setDevServerInPlist(infoPlistPath, devServerURL); err != nil {
-			fmt.Printf("Warning: could not set dev server in Info.plist: %v\n", err)
-		}
-
-		// Start dev server in background
-		fmt.Printf("Starting dev server at %s...\n", devServerURL)
-		devServerCmd = exec.Command("air")
-		devServerCmd.Stdout = os.Stdout
-		devServerCmd.Stderr = os.Stderr
-		if err := devServerCmd.Start(); err != nil {
-			return fmt.Errorf("failed to start dev server: %w", err)
-		}
-
-		// Give server time to start
-		fmt.Println("Waiting for dev server to start...")
-		exec.Command("sleep", "3").Run()
-
-	} else {
-		// Production mode: build the framework
-		modulePath, err := getModulePath()
-		if err != nil {
-			return fmt.Errorf("could not determine module path: %w", err)
-		}
-
-		fmt.Println("Building iOS framework...")
-		if err := buildIOS(modulePath); err != nil {
-			return err
-		}
-
-		// Clear dev server from Info.plist for production builds
-		infoPlistPath := filepath.Join(iosProjectPath, "Example/Info.plist")
-		clearDevServerInPlist(infoPlistPath)
-	}
-
-	appPath, err := buildSimulatorApp(iosProjectPath)
-	if err != nil {
-		if devServerCmd != nil {
-			devServerCmd.Process.Kill()
-		}
-		return err
-	}
-
-	// Find an available iPhone simulator
-	simulatorName := findAvailableIPhoneSimulator()
-	if simulatorName == "" {
-		simulatorName = "iPhone 15" // Fallback
-	}
-
-	// Boot simulator if needed
-	fmt.Printf("Launching iOS Simulator (%s)...\n", simulatorName)
-	runCommand("xcrun", "simctl", "boot", simulatorName) // Ignore error if already booted
-
-	// Open Simulator app
-	runCommand("open", "-a", "Simulator")
-
-	// Install app
-	fmt.Println("Installing app...")
-	if err := runCommand("xcrun", "simctl", "install", "booted", appPath); err != nil {
-		if devServerCmd != nil {
-			devServerCmd.Process.Kill()
-		}
-		return fmt.Errorf("failed to install app: %w", err)
-	}
-
-	// Launch app
-	fmt.Println("Launching app...")
-	bundleID := "com.irgo.Example" // Default bundle ID
-	if err := runCommand("xcrun", "simctl", "launch", "booted", bundleID); err != nil {
-		if devServerCmd != nil {
-			devServerCmd.Process.Kill()
-		}
-		return fmt.Errorf("failed to launch app: %w", err)
-	}
-
-	if devMode {
-		fmt.Println()
-		fmt.Println("===========================================")
-		fmt.Println("iOS app running in DEV MODE with hot reload!")
-		fmt.Printf("Dev server: %s\n", devServerURL)
-		fmt.Println("Edit your Go code and see changes instantly.")
-		fmt.Println("Press Ctrl+C to stop.")
-		fmt.Println("===========================================")
-		fmt.Println()
-
-		// Wait for dev server to exit (user presses Ctrl+C)
-		devServerCmd.Wait()
-	} else {
-		fmt.Println("\nApp running on iOS Simulator!")
-	}
-
-	return nil
-}
-
-// findAvailableIPhoneSimulator finds an available iPhone simulator
-func findAvailableIPhoneSimulator() string {
-	// Get list of available simulators
-	out, err := exec.Command("xcrun", "simctl", "list", "devices", "available", "-j").Output()
-	if err != nil {
-		return ""
-	}
-
-	// Parse JSON to find an iPhone
-	// Look for common iPhone names in priority order
-	preferences := []string{"iPhone 15 Pro", "iPhone 15", "iPhone 17 Pro", "iPhone 17", "iPhone SE"}
-	outStr := string(out)
-	for _, name := range preferences {
-		if strings.Contains(outStr, name) {
-			return name
-		}
-	}
-
-	return ""
-}
-
-func runAndroid(devMode bool, avdName string, headless bool) error {
-	// Self-provision the Android toolchain (JDK/SDK/NDK/gomobile, plus the
-	// emulator + AVD when no device is connected) — devs and CI never need a
-	// separate setup step. Idempotent: only installs what is missing.
-	if err := ensureAndroidToolchain(!adbRunning(), avdName); err != nil {
-		return err
-	}
-
-	// Ensure the canonical android/Example app exists (scaffolded from the
-	// embedded templates when missing — devs and CI never hand-copy it).
-	if err := scaffoldExamples(); err != nil {
-		return fmt.Errorf("Android example scaffold failed: %w", err)
-	}
-
-	// Check if android/Example project exists
-	androidProjectPath := "android/Example"
-	if _, err := os.Stat(androidProjectPath); os.IsNotExist(err) {
-		return fmt.Errorf("Android project not found at %s", androidProjectPath)
-	}
-
-	// App icon: single source icon → launcher mipmaps (if present), so the
-	// installed APK shows the project's icon.
-	if ic := findAppIcon(""); ic != "" {
-		_ = generateAndroidIcons(ic, filepath.Join(androidProjectPath, "app", "src", "main", "res"))
-	}
-
-	// Dev server URL as seen from the emulator (10.0.2.2 is the host loopback)
-	devServerURL := "http://10.0.2.2:8080"
-	var devServerCmd *exec.Cmd
-
-	if devMode {
-		fmt.Println("Running in DEV MODE with hot reload...")
-		fmt.Println()
-
-		// Check for required dev tools
-		if err := ensureGoTool("air"); err != nil {
-			return err
-		}
-
-		// Dev mode serves the app from the dev server, so a fresh gomobile
-		// build isn't needed. The Gradle project still links against
-		// app/libs/irgo.aar, so build it only when missing or built by a
-		// different irgo version (stale bridge API).
-		aarPath := filepath.Join(androidProjectPath, "app/libs/irgo.aar")
-		if !artifactUpToDate(aarPath, "build/android") {
-			modulePath, err := getModulePath()
-			if err != nil {
-				return fmt.Errorf("could not determine module path: %w", err)
-			}
-
-			fmt.Println("Building Android AAR (missing or built by another irgo version)...")
-			if err := buildAndroid(modulePath); err != nil {
-				return err
-			}
-		} else {
-			fmt.Printf("Using existing %s (delete it to force a rebuild)\n", aarPath)
-		}
-
-		// Start dev server in background
-		fmt.Println("Starting dev server at http://localhost:8080...")
-		devServerCmd = exec.Command("air")
-		devServerCmd.Stdout = os.Stdout
-		devServerCmd.Stderr = os.Stderr
-		if err := devServerCmd.Start(); err != nil {
-			return fmt.Errorf("failed to start dev server: %w", err)
-		}
-
-		// Give server time to start
-		fmt.Println("Waiting for dev server to start...")
-		exec.Command("sleep", "3").Run()
-	} else {
-		// Production mode: build the AAR
-		modulePath, err := getModulePath()
-		if err != nil {
-			return fmt.Errorf("could not determine module path: %w", err)
-		}
-
-		fmt.Println("Building Android AAR...")
-		if err := buildAndroid(modulePath); err != nil {
-			return err
-		}
-	}
-
-	killDevServer := func() {
-		if devServerCmd != nil {
-			devServerCmd.Process.Kill()
-		}
-	}
-
-	// Build with Gradle
-	gradlew := filepath.Join(androidProjectPath, "gradlew")
-	if _, err := os.Stat(gradlew); os.IsNotExist(err) {
-		killDevServer()
-		return fmt.Errorf("gradlew not found in %s", androidProjectPath)
-	}
-
-	fmt.Println("Building Android app...")
-	// Run ./gradlew relative to cmd.Dir: exec.Command with a path relative to
-	// the ORIGINAL cwd would not resolve once cmd.Dir switches the working
-	// directory to androidProjectPath.
-	cmd := exec.Command("./gradlew", "assembleDebug")
-	cmd.Dir = androidProjectPath
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		killDevServer()
-		return fmt.Errorf("gradle build failed: %w", err)
-	}
-
-	// Find the built APK
-	apkPath := filepath.Join(androidProjectPath, "app/build/outputs/apk/debug/app-debug.apk")
-	if _, err := os.Stat(apkPath); os.IsNotExist(err) {
-		killDevServer()
-		return fmt.Errorf("built APK not found at %s", apkPath)
-	}
-
-	// Boot the emulator if none is running, then install.
-	if err := ensureEmulatorRunning(avdName, headless); err != nil {
-		killDevServer()
-		return err
-	}
-	fmt.Println("Installing on Android device/emulator...")
-	if err := runCommand(adbBin(), "install", "-r", apkPath); err != nil {
-		killDevServer()
-		return fmt.Errorf("failed to install APK (is an emulator running?): %w", err)
-	}
-
-	// Launch app
-	fmt.Println("Launching app...")
-	packageName := "com.irgo.example"
-	activityName := ".MainActivity"
-	launchArgs := []string{"shell", "am", "start", "-n", packageName + "/" + packageName + activityName}
-	if devMode {
-		// IrgoActivity reads the irgoDevServer extra and loads that URL
-		// instead of the embedded bridge.
-		launchArgs = append(launchArgs, "-e", "irgoDevServer", devServerURL)
-	}
-	if err := runCommand(adbBin(), launchArgs...); err != nil {
-		killDevServer()
-		return fmt.Errorf("failed to launch app: %w", err)
-	}
-
-	if devMode {
-		fmt.Println()
-		fmt.Println("===========================================")
-		fmt.Println("Android app running in DEV MODE with hot reload!")
-		fmt.Printf("Dev server: %s (localhost:8080 on this machine)\n", devServerURL)
-		fmt.Println("Edit your Go code and see changes instantly.")
-		fmt.Println("Press Ctrl+C to stop.")
-		fmt.Println("===========================================")
-		fmt.Println()
-
-		// Wait for dev server to exit (user presses Ctrl+C)
-		devServerCmd.Wait()
-	} else {
-		fmt.Println("\nApp running on Android!")
-	}
-
-	return nil
-}
-
-// Helper functions
-
-func checkTool(name, installCmd string) error {
-	_, err := exec.LookPath(name)
-	if err != nil {
-		return fmt.Errorf("%s not found. Install with: %s", name, installCmd)
-	}
-	return nil
-}
-
-func runCommand(name string, args ...string) error {
-	cmd := exec.Command(name, args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	cmd.Stdin = os.Stdin
-	return cmd.Run()
-}
-
-func getModulePath() (string, error) {
-	// Try to read from go.mod
-	data, err := os.ReadFile("go.mod")
-	if err != nil {
-		return "", err
-	}
-
-	// Parse module line
-	lines := string(data)
-	for _, line := range splitLines(lines) {
-		if len(line) > 7 && line[:7] == "module " {
-			return line[7:], nil
-		}
-	}
-
-	return "", fmt.Errorf("module directive not found in go.mod")
-}
-
-func splitLines(s string) []string {
-	var lines []string
-	start := 0
-	for i := 0; i < len(s); i++ {
-		if s[i] == '\n' {
-			line := s[start:i]
-			if len(line) > 0 && line[len(line)-1] == '\r' {
-				line = line[:len(line)-1]
-			}
-			lines = append(lines, line)
-			start = i + 1
-		}
-	}
-	if start < len(s) {
-		lines = append(lines, s[start:])
-	}
-	return lines
-}
-
-func copyFile(src, dst string) error {
-	data, err := os.ReadFile(src)
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(dst, data, 0644)
-}
-
-// ensureMobileBuildSetup ensures the go.work file and x/mobile are set up correctly
-func ensureMobileBuildSetup() error {
-	goVersion := getGoVersion()
-
-	// Check if go.work exists with x/mobile
-	if _, err := os.Stat("go.work"); os.IsNotExist(err) {
-		// Get irgo path for replacement
-		irgoPath := getIrgoPath()
-
-		// Clone x/mobile if not already present
-		mobileDir := filepath.Join(os.TempDir(), "golang-mobile")
-		if _, err := os.Stat(mobileDir); os.IsNotExist(err) {
-			fmt.Println("Cloning golang.org/x/mobile...")
-			cmd := exec.Command("git", "clone", "--depth", "1", "https://github.com/golang/mobile", mobileDir)
-			cmd.Stdout = os.Stdout
-			cmd.Stderr = os.Stderr
-			if err := cmd.Run(); err != nil {
-				return fmt.Errorf("failed to clone x/mobile: %w", err)
-			}
-		}
-
-		// Update go.mod in cloned repo to use current Go version
-		mobileModPath := filepath.Join(mobileDir, "go.mod")
-		if data, err := os.ReadFile(mobileModPath); err == nil {
-			content := string(data)
-			// Replace any go 1.x.x version with current version
-			lines := splitLines(content)
-			for i, line := range lines {
-				if len(line) > 3 && line[:3] == "go " {
-					lines[i] = "go " + goVersion
-					break
-				}
-			}
-			os.WriteFile(mobileModPath, []byte(strings.Join(lines, "\n")), 0644)
-		}
-
-		// Create go.work file
-		workContent := fmt.Sprintf("go %s\n\nuse (\n\t.\n", goVersion)
-		if irgoPath != "" {
-			workContent += fmt.Sprintf("\t%s\n", irgoPath)
-		}
-		workContent += fmt.Sprintf("\t%s\n)\n", mobileDir)
-
-		if err := os.WriteFile("go.work", []byte(workContent), 0644); err != nil {
-			return fmt.Errorf("failed to create go.work: %w", err)
-		}
-		fmt.Println("Created go.work for mobile build")
-
-		// Install gomobile and gobind from local source
-		fmt.Println("Installing gomobile from source...")
-		cmd := exec.Command(goBin(), "install", "./cmd/gomobile")
-		cmd.Dir = mobileDir
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to install gomobile: %w", err)
-		}
-
-		cmd = exec.Command(goBin(), "install", "./cmd/gobind")
-		cmd.Dir = mobileDir
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to install gobind: %w", err)
-		}
-
-		// `go install` lands in GOBIN (default $GOPATH/bin), which may not be on
-		// the caller's PATH — prepend it so runGomobileCommand resolves gomobile.
-		_ = os.Setenv("PATH", gobinDir()+string(os.PathListSeparator)+os.Getenv("PATH"))
-	}
-
-	return nil
-}
-
-// runGomobileCommand runs a gomobile command with the correct GOTOOLCHAIN
-func runGomobileCommand(args ...string) error {
-	goVersion := getGoVersion()
-
-	// gomobile bind for Android shells out to javac, which needs a JDK on
-	// PATH — resolve the managed ~/.irgo/jdks JDK (or an existing one) so the
-	// toolchain is self-contained after `irgo install-tools android`.
-	applyBestJDKToEnv()
-
-	cmd := exec.Command("gomobile", args...)
-	cmd.Env = append(os.Environ(), "GOTOOLCHAIN=go"+goVersion)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	cmd.Stdin = os.Stdin
-	return cmd.Run()
-}
-
-// setDevServerInPlist adds IRGO_DEV_SERVER to Info.plist
-func setDevServerInPlist(plistPath, devServerURL string) error {
-	data, err := os.ReadFile(plistPath)
-	if err != nil {
-		return err
-	}
-
-	content := string(data)
-
-	// Check if IRGO_DEV_SERVER already exists
-	if strings.Contains(content, "IRGO_DEV_SERVER") {
-		// Update existing value
-		// Find and replace the string value after IRGO_DEV_SERVER key
-		start := strings.Index(content, "<key>IRGO_DEV_SERVER</key>")
-		if start != -1 {
-			stringStart := strings.Index(content[start:], "<string>")
-			stringEnd := strings.Index(content[start:], "</string>")
-			if stringStart != -1 && stringEnd != -1 {
-				newContent := content[:start+stringStart+8] + devServerURL + content[start+stringEnd:]
-				return os.WriteFile(plistPath, []byte(newContent), 0644)
-			}
-		}
-	}
-
-	// Add new key-value pair before </dict>
-	insertPoint := strings.LastIndex(content, "</dict>")
-	if insertPoint == -1 {
-		return fmt.Errorf("could not find </dict> in Info.plist")
-	}
-
-	newEntry := fmt.Sprintf("\t<key>IRGO_DEV_SERVER</key>\n\t<string>%s</string>\n", devServerURL)
-	newContent := content[:insertPoint] + newEntry + content[insertPoint:]
-
-	return os.WriteFile(plistPath, []byte(newContent), 0644)
-}
-
-// clearDevServerInPlist removes IRGO_DEV_SERVER from Info.plist
-func clearDevServerInPlist(plistPath string) {
-	data, err := os.ReadFile(plistPath)
-	if err != nil {
-		return
-	}
-
-	content := string(data)
-
-	// Find and remove IRGO_DEV_SERVER entry
-	keyStart := strings.Index(content, "<key>IRGO_DEV_SERVER</key>")
-	if keyStart == -1 {
-		return
-	}
-
-	// Find the end of the string value
-	valueEnd := strings.Index(content[keyStart:], "</string>")
-	if valueEnd == -1 {
-		return
-	}
-
-	// Remove the entire entry including newlines
-	entryEnd := keyStart + valueEnd + len("</string>")
-
-	// Look for trailing newline
-	if entryEnd < len(content) && content[entryEnd] == '\n' {
-		entryEnd++
-	}
-
-	// Look for leading whitespace/newline
-	entryStart := keyStart
-	for entryStart > 0 && (content[entryStart-1] == '\t' || content[entryStart-1] == ' ') {
-		entryStart--
-	}
-
-	newContent := content[:entryStart] + content[entryEnd:]
-	os.WriteFile(plistPath, []byte(newContent), 0644)
-}
-
-// uninstallTools is the exact inverse of `irgo install-tools`: it removes the
-// Go tools irgo installed, and nothing else. Every install path in the CLI has
-// a matching uninstall — without one you cannot return a machine to a known
-// state, and a provisioning bug hides behind whatever was left lying around
-// instead of surfacing on the next run.
-//
-// Marker-guarded: a tool irgo did not install is reported and kept, so a
-// developer's own templ/air survives. Pass all to override that.
-func uninstallTools(all bool) error {
-	fmt.Println("Removing irgo-installed Go tools...")
-
-	// A tool can sit in the resolved GOBIN and/or the default $GOPATH/bin —
-	// they differ when GOBIN was empty at install time. Check both.
-	binDirs := []string{gobinDir()}
-	if out, err := exec.Command(goBin(), "env", "GOPATH").Output(); err == nil {
-		if gp := strings.TrimSpace(string(out)); gp != "" {
-			if p := filepath.Join(gp, "bin"); p != binDirs[0] {
-				binDirs = append(binDirs, p)
-			}
-		}
-	}
-
-	removed, kept, missing := 0, 0, 0
-	for _, tool := range []string{"templ", "air", "gomobile", "gobind"} {
-		if !all && !toolInstalledByIrgo(tool) {
-			if _, err := exec.LookPath(tool); err == nil {
-				fmt.Printf("  %s: kept (not installed by irgo — use --all to remove anyway)\n", tool)
-				kept++
-			} else {
-				missing++
-			}
-			continue
-		}
-		found := false
-		for _, dir := range binDirs {
-			name := tool
-			if runtime.GOOS == "windows" {
-				name += ".exe"
-			}
-			p := filepath.Join(dir, name)
-			if _, err := os.Stat(p); err == nil {
-				if err := os.Remove(p); err != nil {
-					return fmt.Errorf("removing %s: %w", p, err)
-				}
-				fmt.Printf("  %s: removed (%s)\n", tool, p)
-				found = true
-				removed++
-			}
-		}
-		if !found {
-			missing++
-		}
-		_ = os.Remove(filepath.Join(irgoToolsDir(), tool))
-	}
-
-	// Build residue from mobile builds: the temp x/mobile clone and the local
-	// go.work that ensureMobileBuildSetup generates.
-	_ = os.RemoveAll(filepath.Join(os.TempDir(), "golang-mobile"))
-	_ = os.Remove("go.work")
-	_ = os.Remove("go.work.sum")
-
-	if err := uninstallMinGW(all); err != nil {
-		return err
-	}
-
-	fmt.Printf("\n%d removed, %d kept, %d not present.\n", removed, kept, missing)
-	fmt.Println("Android SDK/NDK/JDK are separate: irgo uninstall-tools android --remove-jdk")
-	return nil
-}
-
-// uninstallMinGW removes the mingw-w64 cross-compiler, but only when irgo
-// installed it — ensureMinGW brew-installs it on macOS for Windows builds.
-func uninstallMinGW(all bool) error {
-	if runtime.GOOS != "darwin" {
-		return nil
-	}
-	if _, err := exec.LookPath("x86_64-w64-mingw32-gcc"); err != nil {
-		return nil
-	}
-	if !all && !toolInstalledByIrgo("mingw-w64") {
-		fmt.Println("  mingw-w64: kept (not installed by irgo — use --all to remove anyway)")
-		return nil
-	}
-	if _, err := exec.LookPath("brew"); err != nil {
-		fmt.Println("  mingw-w64: present but Homebrew is unavailable — remove it manually")
-		return nil
-	}
-	fmt.Println("  mingw-w64: removing via brew...")
-	if err := runCommand("brew", "uninstall", "--formula", "mingw-w64"); err != nil {
-		fmt.Printf("  Warning: brew uninstall mingw-w64 failed: %v\n", err)
-	}
-	_ = os.Remove(filepath.Join(irgoToolsDir(), "mingw-w64"))
-	return nil
 }
